JWT OOP logic in helpers
